Extract onMount & events to separate modules
We've slowly cut down the amount of code actually in the component.
This is the bare minimum in the component, with onMount & events
logic split off, making it eaiser to reason about each piece.
