+++ client/editor/instance/onMount.js
@@ -1,0 +1,231 @@
+import R from 'ramda';
+import { concat, fromEvents, fromPromise, merge, never, stream } from 'kefir';
+import Prism from '../../prism';
+import render from 'brookjs/render';
+import template from './index.hbs';
+
+const renderTemplate = render(template);
+
+// const CRLF = /\r?\n|\r/g;
+
+/**
+ * Update the highlighted line numbers next to the editor.
+ *
+ * @returns {Observable} Observable to update the editor line numbers.
+ */
+function updateLinenumber(/*pre, start, end*/) {
+    return stream(emitter => {
+        // let content = pre.textContent;
+        // let ss = pre.selectionStart;
+        // let se = pre.selectionEnd;
+        //
+        // // @todo push into store
+        // ss && pre.setAttribute('data-ss', ss);
+        // se && pre.setAttribute('data-se', se);
+
+        // Update current line highlight
+        // let line = (content.slice(0, ss).match(CRLF) || []).length;
+
+        // pre.setAttribute('data-line', line + 1);
+
+        emitter.end();
+    });
+}
+
+/**
+ * Resets the cursor's selection range.
+ *
+ * @param {Element} node - Node to add selection range to.
+ * @param {number} ss - Selection start.
+ * @param {number} se - Selection end.
+ * @returns {Observable} Observable to update selection range.
+ */
+const setSelectionRange = R.curry(function setSelectionRange(node, ss, se) {
+    return stream(emitter => {
+        const loop = requestAnimationFrame(() => {
+            let range = document.createRange();
+            let offset = findOffset(node, ss);
+
+            range.setStart(offset.element, offset.offset);
+
+            if (se && se !== ss) {
+                offset = findOffset(node, se);
+            }
+
+            range.setEnd(offset.element, offset.offset);
+
+            let selection = window.getSelection();
+            selection.removeAllRanges();
+            selection.addRange(range);
+
+            emitter.end();
+        });
+
+        return () => cancelAnimationFrame(loop);
+    });
+});
+
+/**
+ * Find the offset for a given selection start
+ *
+ * @param {Element} root - Element to find offset for.
+ * @param {number} ss - Start point for offset.
+ * @returns {Object} Offset information.
+ */
+function findOffset(root, ss) {
+    let container;
+
+    if (!root) {
+        return null;
+    }
+
+    let offset = 0;
+    let element = root;
+
+    do {
+        container = element;
+        element = element.firstChild;
+
+        if (element) {
+            do {
+                let len = element.textContent.length;
+
+                if (offset <= ss && offset + len > ss) {
+                    break;
+                }
+
+                offset += len;
+            } while (element = element.nextSibling);
+        }
+
+        if (!element) {
+            // It's the container's lastChild
+            break;
+        }
+    } while (element && element.hasChildNodes() && element.nodeType !== 3);
+
+    if (element) {
+        return {
+            element: element,
+            offset: ss - offset
+        };
+    } else if (container) {
+        element = container;
+
+        while (element && element.lastChild) {
+            element = element.lastChild;
+        }
+
+        if (element.nodeType === 3) {
+            return {
+                element: element,
+                offset: element.textContent.length
+            };
+        } else {
+            return {
+                element: element,
+                offset: 0
+            };
+        }
+    }
+
+    return {
+        element: root,
+        offset: 0,
+        error: true
+    };
+}
+
+/**
+ * Highlight the given element with Prism.
+ *
+ * @param {Element} el - Element to highlight.
+ * @returns {Observable} Observable to highlight element.
+ */
+const highlightElement = function highlightElement(el) {
+    return stream(emitter => {
+        const loop = requestAnimationFrame(() => {
+            Prism.highlightElement(el.querySelector('code'), false);
+            emitter.end();
+        });
+
+        return () => cancelAnimationFrame(loop);
+    });
+};
+
+/**
+ * Create a debounced keyup stream.
+ *
+ * @param {Element} el - Element to create stream for.
+ * @returns {Observable} Observable of keyup events.
+ */
+const createDebouncedKeyup = function createDebouncedKeyup(el) {
+    return fromEvents(el, 'keyup').debounce(150);
+};
+
+/**
+ * Creates a stream to update Prism's setting and fetch any plugin scripts.
+ *
+ * @param {Object} state - Current page state.
+ * @returns {Observable} Observable that ends after Prism is updated.
+ */
+const createPrismUpdateStream = function createPrismUpdateStream(state) {
+    return fromPromise(Promise.all([
+        Prism.setTheme(state.editor.theme),
+        Prism.togglePlugin('show-invisibles', state.editor.invisibles === 'on')
+    ]))
+        .ignoreValues();
+};
+
+/**
+ * Creates a stream to update the element after state updates.
+ *
+ * @param {Element} el - Element to render.
+ * @param {Object} state - Updated state.
+ * @returns {Observable} Observable to update element.
+ */
+const createDOMUpdateStream = R.curry(function createDOMUpdateStream(el, state) {
+    return merge([
+        renderTemplate(el, state, state),
+        highlightElement(el),
+        state.instance.cursor ? setSelectionRange(el.querySelector('code'), ...state.instance.cursor) : never(),
+        state.instance.cursor ? updateLinenumber(el.querySelector('pre'), ...state.instance.cursor) : never()
+    ]);
+});
+
+/**
+ * Creates a stream to update Prism's settings then update the DOM.
+ *
+ * @param {Element} el - Element to update.
+ * @param {Element} state - Page state to update to.
+ * @returns {Observable} Observable to render the element.
+ */
+const createRenderStream = R.curry(function createRenderStream(el, state) {
+    return concat([createPrismUpdateStream(state), createDOMUpdateStream(el, state)]);
+});
+
+/**
+ * Creates a new stream on mount that handles updating Prism & rendering the editor.
+ *
+ * This gives us finer control over the render cycle than merely deferring to `brookjs`,
+ * as we don't just want to rerender on state changes but when the keyup event occurs,
+ * allowing us to rerender the highlighting between typings, rather than interrupting
+ * the user.
+ *
+ * @param {Element} el - Editor element.
+ * @param {Observable} props$ - Stream of editor props.
+ * @returns {Observable} - Stream of renders.
+ */
+export default R.curry(function onMount(el, props$) {
+    const setAutoloader$ = stream(emitter => {
+        Prism.setAutoloaderPath(__webpack_public_path__);
+
+        emitter.end();
+    });
+
+    const render$ = props$.sampledBy(createDebouncedKeyup(el))
+        .merge(props$.take(1))
+        .flatMapLatest(createRenderStream(el));
+
+    return merge([setAutoloader$, render$]);
+});
